+++ ResultFunctionalXUnit/FunctionalExtensions/Sync/ResultExtension/ResultCollectionTest/ToResultCollectionExtensionsTest.cs
@@ -43,39 +43,5 @@ namespace ResultFunctionalXUnit.FunctionalExtensions.Sync.ResultExtension.Result
             Assert.True(result.HasErrors);
             Assert.True(result.Errors.First().Equals(errorInitial));
         }
-
-        /// <summary>
-        /// Проверить объект на нул. Ошибка нулевого значения
-        /// </summary>
-        [Fact]
-        public void ConcatResultCollection_Ok()
-        {
-            var resultCollectionFirst = GetRangeNumber().ToResultCollection();
-            var resultCollectionSecond = GetRangeNumber().ToResultCollection();
-            var results = Enumerable.Empty<IResultCollection<int>>().Append(resultCollectionFirst).Append(resultCollectionSecond);
-
-            var resultCollection = results.ConcatResultCollection();
-            var numberRange = resultCollectionFirst.Value.Concat(resultCollectionSecond.Value).ToList();
-
-            Assert.True(resultCollection.OkStatus);
-            Assert.True(numberRange.SequenceEqual(resultCollection.Value));
-        }
-
-        /// <summary>
-        /// Проверить объект на нул. Ошибка нулевого значения
-        /// </summary>
-        [Fact]
-        public void ConcatResultCollection_Error()
-        {
-            var resultCollectionFirst = GetRangeNumber().ToResultCollection();
-            var resultCollectionErrorFirst = CreateErrorTest().ToResultCollection<int>();
-            var resultCollectionErrorSecond = CreateErrorTest().ToResultCollection<int>();
-            var results = Enumerable.Empty<IResultCollection<int>>().Append(resultCollectionFirst).Append(resultCollectionErrorFirst).Append(resultCollectionErrorSecond);
-
-            var resultCollection = results.ConcatResultCollection();
-
-            Assert.True(resultCollection.HasErrors);
-            Assert.Equal(2, resultCollection.Errors.Count);
-        }
     }
 }
